pyproject.toml
[project]
name = "dnd-summary"
version = "0.0.0"
description = "Temporal-orchestrated D&D transcript summarization pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "temporalio>=1.7",
  "typer>=0.12",
  "python-docx>=1.1",
  "google-genai>=1.0.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "python-multipart>=0.0.9",
  "pgvector>=0.3.6",
]

[project.optional-dependencies]
dev = [
  "black>=24.3",
  "alembic>=1.13",
  "flake8>=7.0",
  "pytest>=8.2",
]
eval = [
  "dspy-ai>=2.4.0",
]

[project.scripts]
dnd-summary = "dnd_summary.cli:app"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
